Separate reactive primitives into own package.
This package contains the core reactive primitives
and logic used by the bobtail framework. They have
been separated them into their own package so that
they are no longer tightly coupled with the
DOM templating framework that ships with Bobtail.

This is primarily useful for those using Javascript
in a server environment--i.e., NodeJS--as they can
now use the bobtail primitives without the templating
code (useless on the backend), and also without
the browser specific logic used in bobtail.

Furthermore, this allows us to develop alternate
templating systems if we choose to do so in the future.

This also allows us to remove our dependencies on
JQuery and es5/6-shim.

Git history from bobtail has been maintained, but tags
have been removed.