Fix https://github.com/bobtail-dev/bobtail-rx/issues/13
Previously, when refreshing a cell, we would compute all
of its downstream dependencies, and refresh them in topological
order at most once. However, when refreshing a cell, we recompute its
dependencies. That means dependencies can change during the refresh
process, requiring a cell to be refreshed twice. This can leave us
in an inconsistent state.

This patch fixes this problem. Rather than iterating over the
topologically sorted list, we use a hybrid set-queue structure.
We consume cells to refresh from the front of the queue, and
append its dependencies to the back. However, because we're using
a Set to maintain our queue, any dependencies which are already
in the queue won't be re-appended. Cells are also unshielded
immediately after they refresh, rather than at the end when
everything has been refreshed.

This ensures that if the dependency structure changes and we
end up refreshing out of topological order, we will re-refresh
any cells that need to be, and end in a consistent state.